# Performance improvement of `FermionicOp.reduce`