<!--
⚠️ If you do not respect this template, your pull request will be closed.
⚠️ Your pull request title should be short detailed and understandable for all.
⚠️ Also, please add it in the CHANGELOG file under Unreleased section.
⚠️ If your pull request fixes an open issue, please link to the issue.

✅ I have added the tests to cover my changes.
✅ I have updated the documentation accordingly.
✅ I have read the CONTRIBUTING document.
-->

### Summary

This PR introduces the following two changes.
- Vectorize `FerimionicOp.reduce`
- Add `normal_order` option to `FerimionicOp.reduce` (default is True)

```python
import random
from timeit import timeit
from qiskit_nature.operators.second_quantization import FermionicOp


def random_op(k, n):
    random.seed(123)
    op = sum((FermionicOp("".join(random.choices("+-ENI", k=k)), display_format="dense") * random.random()) for _ in range(n))
    return op

def bench_reduce(op):
    print(f"normal order: {timeit(lambda: op.reduce(), number=1)} sec")
    try:
        print(f"no normal order: {timeit(lambda: op.reduce(normal_order=False), number=1)} sec")
    except TypeError:
        pass

op = random_op(5, 10000)
bench_reduce(op)
```

main
```
normal order: 2.4304106099998535 sec
```

this PR
```
normal order: 2.271573748999799 sec
no normal order: 0.05692009299946221 sec
```

### Details and comments

The motivation to apply `reduce` before qubit mapping is #585. The output operator can be redundant.
So, I want to simplify the operators. But, I don't think qubit operator requires the normal order.

I added the `normal_order` option because applying `normal_order` may increase the number of operators and `to_normal_order` can be slow.
If users apply `reduce` to an operator before qubit mapping, it takes more time than without `reduce` due to the larger number of operators.
I attach an example. The size of operator grows from 1000 to 18950 by applying `reduce`.

```python
import random
from timeit import timeit

from qiskit_nature.operators.second_quantization import FermionicOp
from qiskit_nature.converters.second_quantization import QubitConverter
from qiskit_nature.mappers.second_quantization import JordanWignerMapper

def random_op(k, n):
    random.seed(123)
    op = sum((FermionicOp("".join(random.choices("+-E", k=k)), display_format="dense") * random.random()) for _ in range(n))
    return op

qubit_converter = QubitConverter(mapper=JordanWignerMapper())

def bench_convert(op):
    print(f"w/o reduce: {timeit(lambda: qubit_converter.convert(op), number=1)} sec")
    print(f"w/ reduce of normal order: {timeit(lambda: qubit_converter.convert(op.reduce()), number=1)} sec")
    try:
        print(f"w/ reduce no normal order: {timeit(lambda: qubit_converter.convert(op.reduce(normal_order=False)), number=1)} sec")
    except TypeError:
        pass

op = random_op(10, 1000)

print("qubit convert")
bench_convert(op)

print("\noperator sizes")
print(f"original op: {len(op)}")
print(f"w/ reduce of normal order: {len(op.reduce())}")
try:
    print(f"w/ reduce no normal order: {len(op.reduce(normal_order=False))}")
except TypeError:
    pass
```

this PR
```
qubit convert
w/o reduce: 1.7863747950004836 sec
w/ reduce of normal order: 17.12350188399978 sec
w/ reduce no normal order: 1.741173968000112 sec

operator sizes
original op: 1000
w/ reduce of normal order: 18950
w/ reduce no normal order: 989
```