@t-imamichi 2022-03-07
I have not written the docstring of the new option `reduce(normal_order)`. If the option is allowed, I will write it.