@t-imamichi 2022-03-23
I will port this optimization to the upcoming `simplify` methods #607. 

https://github.com/Qiskit/qiskit-nature/pull/607#issuecomment-1076457920